Refactoring to breaking into service/model package strategy
