+++ gtfs-validation-lib/src/main/java/com/conveyal/gtfs/model/DuplicateStops.java
@@ -1,4 +1,4 @@
-package com.conveyal.validation;
+package com.conveyal.gtfs.model;
 
 import org.onebusaway.gtfs.model.Stop;
