+++ gtfs-validation-lib/src/main/java/com/conveyal/gtfs/model/Statistic.java
@@ -1,4 +1,4 @@
-package com.conveyal.gtfs;
+package com.conveyal.gtfs.model;
 
 import java.util.Date;
